+++ packages/rax-scrollview/index.d.ts
@@ -1,0 +1,5 @@
+declare namespace JSX {
+  interface IntrinsicElements {
+    'scroll-view': any;
+  }
+}
