+++ packages/rax-scrollview/src/miniapp-native/miniapp/index.axml
@@ -1,13 +1,13 @@
 <scroll-view
     class="rax-scrollview {{className}}"
     style="flex-direction:{{horizontal?'row':'column'}};{{style}}"
-    scroll-x="{{horizontal}}"
-    scroll-y="{{!horizontal}}"
-    scroll-top="{{scrollTop||scroll_top}}"
-    scroll-left="{{scrollLeft||scroll_left}}"
-    scroll-into-view="{{scroll_into_view}}"
-    scroll-with-animation="{{scroll_with_animation}}"
-    scroll-animation-duration="{{scroll_animation_duration}}"
+    scroll-x="{{!disableScroll&&horizontal}}"
+    scroll-y="{{!disableScroll&&!horizontal}}"
+    scroll-top="{{scrollTop||scrollY}}"
+    scroll-left="{{scrollLeft||scrollX}}"
+    scroll-into-view="{{scrollIntoViewId}}"
+    scroll-with-animation="{{scrollWithAnimation}}"
+    scroll-animation-duration="{{scrollAnimationDuration}}"
     lower-threshold="{{endReachedThreshold}}"
     onScroll="onScroll"
     onScrollToLower="onEndReached"
