+++ packages/rax-scrollview/src/miniapp-native/wechat-miniprogram/fmtEvent.ts
