+++ packages/rax-scrollview/src/miniapp-native/wechat-miniprogram/index.wxml
@@ -1,0 +1,16 @@
+<scroll-view
+    class="rax-scrollview {{className}}"
+    style="flex-direction:{{horizontal?'row':'column'}};{{styleSheet}}"
+    scroll-x="{{!disableScroll&&horizontal}}"
+    scroll-y="{{!disableScroll&&!horizontal}}"
+    scroll-top="{{scrollTop||scrollY}}"
+    scroll-left="{{scrollLeft||scrollX}}"
+    scroll-into-view="{{scrollIntoViewId}}"
+    scroll-with-animation="{{scrollWithAnimation}}"
+    scroll-animation-duration="{{scrollAnimationDuration}}"
+    lower-threshold="{{endReachedThreshold}}"
+    onScroll="onScroll"
+    onScrollToLower="onEndReached"
+>
+    <slot/>
+</scroll-view>
