--- packages/rax-scrollview/src/miniapp-wechat/index.wxml
@@ -1,17 +1,0 @@
-<scroll-view
-    class="rax-scrollview {{className}}"
-    style="flex-direction:{{horizontal?'row':'column'}};{{styleSheet}}"
-    scroll-x="{{horizontal}}"
-    scroll-y="{{!horizontal}}"
-    scroll-top="{{scrollTop||scroll_top}}"
-    scroll-left="{{scrollLeft||scroll_left}}"
-    scroll-into-view="{{scroll_into_view}}"
-    scroll-with-animation="{{scroll_with_animation}}"
-    scroll-animation-duration="{{scroll_animation_duration}}"
-    lower-threshold="{{endReachedThreshold}}"
-    onScroll="onScroll"
-    onScrollToLower="onEndReached"
->
-    <slot/>
-</scroll-view>
-
