+++ packages/rax-scrollview/tsconfig.json
@@ -19,10 +19,10 @@
     "rootDir": "src",
     "baseUrl": ".",
     "paths": {
-      "rax": ["node_modules/@rax-types/rax", "../../node_modules/@rax-types/rax"]
+      "rax": ["node_modules/@rax-types/rax", "node_modules/@rax-types/rax", "index.d.ts"]
     }
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "index.d.ts"],
   "exclude": [
     "node_modules",
     "lib"
